refactor: Change the flow of the application
The idea from arnau:

At this point in the flow, you know you received an event from the Slack
API. Just with that you can do the main split between the two actions
your bot is able to do: Deal with a URL in an event (type message) or
deal with a reaction in an event (type reaction_added).

All of that is entangled in this parse function, to the extent where
both data structures are intermingled. I think we can do better. To do
that, I will ignore for a moment the current implementation and dump my
thoughts on how the flow should go. I'll try to keep the naming as it is
right now to keep things tangible (hopefully :D). So,

A) The entry point (main). It has one-ish responsibility: gather config
from the environment and start the event loop. Let's say it's the
initialisation process. It keeps track of the cache for URL and
confirmation bits but we can ignore this for now.

The main loop delegates to keep_wanted_urls for the event consumption.

B) The event consumer. It has to detect known event types and fire a
specialised process to parse each of them. This could be something on
the lines of:

def event_consumer(events):
    # ...
    for event in events:
    # ...
      if event['type'] == 'message':
         message = parse_message(event)
      if event['type'] == 'reaction_added':
          reaction = parse_reaction_added(event)

I would say, this is the right place to handle the confirmation (I might
change my mind in further iterations of the code) so, there should be
some sort of control for dealing with confirmation. For example:

if confirmation:
    if event['type'] == 'reaction_added':
        return parse_reaction_added(event)

C) Event handler. parsers are specialised by event type so this step has
multiple representations (as many as handled events). If I'm not
mistaken, the handlers have no need for knowing about the confirmation
process.

So, with that in mind, A knows about the initialisation (and the cache),
B knows about the event types and composes the cache as a result of Cs +
confirmation.